Add offline states to the new playlist screen (#7351)
Enable NEW_PLAYLIST_SCREEN

* Fix shuffle test on playlists.

The presenter did not terminate the loop after applying an action on the last view model. This opened the window to inconstant behavior or infinite loops.

* Fix offline status update in the new playlist detail screen. 

A race confusion may occur and report the wrong state. User a map over the last ViewModel when notifying to the view to prevent it.

* Fix crash when making a playlist offline if the server sends back an error.

* Fix missing impression

* Disabled deprecated test