Remove smooth scroll when the user taps the new items pill (#7466)
This change aims to prevent a crash (java.lang.IndexOutOfBoundsException: Inconsistency detected. Invalid view holder adapter ...) caused by a race condition between the smooth scrolling and the adapter updates triggered by the binding being rebuilt. Infact once rebindBinding(null).connect() is called the underlying datasource emits, we clear the adapter and then we add the new items to it. This obviously fails if we smooth scroll a long list because in that way we basically clear the adapter during the list is scrolled.