Fix crash at startup #7090  (#7498)
* Fix crash at startup #7090
On fresh install of a debug build, the app will crash a few seconds after leak canary shows up a memory leak.

The crash is "BadParcelableException: ClassNotFoundException when unmarshalling: SyncJobResult"

It is linked to leak canary and will only affect debug builds. When leak canary spawns a thread to show its leak, somehow it hooks onto our ResultReceiver and returns our bundle but with a null class loader. We have an option to do
resultData.setClassLoader(SyncJobResult.class.getClassLoader())
in the ResultReceiverAdapter but it feels a little hacky.

Another option is to modify the leak canary set-up so that it doesn't initialise the application if we are in an analyser process, which is what the docs recommend.

This commit removes the DevToolsHelper and initialises directly in the application class if we're not in Analyzer Process.

There are two root causes of the issue
1. The app should not try to sync if you're not logged in, which we will need to resolve
2. There is a memory leak in the login page. We're going to leave this in for now as the screen will be re-written