Fix: The playlist detail screen presents a call to action to download the playlist when it is actually already downloaded. (#7575)
It happened when toggling mark/unmark offline. Due to an unsubscribe to the RX call chain, it was possible the store a state w/o publishing it. This operation is now atomic from an RX perspective. 

This issue became noticeable on a Nexus 5 with release soundcloud-android-2017.01.24-557-a390f57-release-586 (commit 39de3dd) but I believe it’s been always there.