Fix crash when playing a track from a position in the playlist. (#7638)
* This could not be reproduced but was caught on production by fabric. The hypothesis is that the track list is updated just after the user clicked on an item (for instance the play count changed because the syncer kicked in), hence the track is not found and the app fails to start playback from position -1. 

The fix consists of matching on the URN (rather than the track item) to mitigate the moving parts - assuming the track urn should be stable (as opposed to properties like play count).  

Note: We had a similar issue for as a long as I remember on other screens and the issue was mitigated by fixing the position (“adjustPositions” helpers) in the PlaybackInitiator. If it does not work we will have to go with a similar solution. 


https://fabric.io/soundcloudandroid/android/apps/com.soundcloud.android/issues/57d91f9f0aeb16625b3a2ac1?time=last-thirty-days

* fix missing throw exception