Make SelectionItem urn Optional to match the backend (#7771)
This required some changes to ImageOperations to allow creation of image placeholders without Urn. The Urn was only used to create a cache-key, so I added new methods for the caller to just provide their own cache-key directly. We should probably consider re-jigging all of this code at some point as it's becoming a bit of a maze