Override origin page name when coming from Recently Played bucket (#7773)
* Introduce the `PlaySessionOriginScreenProvider` to provide the origin screen of a play session source.

This could be either: 
- the current screen
- the collection screen (to track the recently played bucket)

* Subscribe the PlaySessionOriginScreenProvider when the app starts

* Fix build

* Draft implementation of the `PlaySessionOriginScreenProvider`

- Add tests
- Partial implementation

* Override origin page name when coming from Recently Played bucket

ScreenProvider, the default mechanism for tracking page views, updates its value for every onResume
that is called on every fragment of the app. In order to track the origin of the “plays” that come from the
recently played bucket, we used to be able to follow that mechanism.

This changed when we introduced intermediary screens for Stations, Profiles and Playlists. Now, clicking
one of those entities on the Recently Played bucket will lead the user to a middle page that will eventually
override the value stored in ScreenProvider.

This change overrides that behaviour, since we want to track those plays as coming from the Collection
screen instead. This means that the following user flow:
Collection —> Click on recently played item (station/profile/playlist) —> intermediary screen —> play
should be counted as a play in the Collection page instead of the intermediary screen.

This mechanism is temporary and should be reverted when we implement Holistic tracking on the app,
in a way that we could add references from events to other events, in a way to say that “a play has a parent
event that is a page view on the collection screen” and attribute it that way. Since this project is still far from
being implemented, we have to live with this workaround for now.

* Rollback changes on ProfileHeaderPresenter

We want to override the Screen tag for playing tracks instead of following a user.
This reverts the logic to track “user following” with the
data from PlaySessionOriginScreenProvider.

* Override profile tracks playing to send collection screen tag

* Remove redundant initialization

* Fix build

* Revert alterations on playlist screen