Remove onDurationChange callback handling (#7935)
When transitioning from the Preparing to Prepared states, Flipper will
fetch the HLS duration from the backend. This might differ by a few
milliseconds from the value we receive from API-mobile (and thus store
in the database) in normal use cases.

This value was only being used when reporting the progress while seeking.

We should always trust the state the player sends back to us. That means:
`state_change` objects are the representation of those states and from there we
should take out the duration values via `getDuration()`.

This removes the state holding for the `duration` attribute and removes the double
progress reporting while seeking. There should be no visible change to the user, except
that we could even have a smoother waveform animation, since a single progress report
will be done for each update we receive from the player.