Support Latin American Spanish (#7974)
For more context, see the comment on the ticket I18N-299

Transifex will send Latin American Spanish text through in the 'es_419' directory. We then map es_419 to b+es+419 (which is the correct format), but then also copy that file into values-es-rUS directory to ensure backwards compatibility

Transifex does not support one to many mappings so unfortunately this is a bit hacky. I will update the wiki as we will no longer be able to use transifex pull or the file will not be copied to the es/US folder. We will have to run this script instead.

Side note, when testing this I realised that the lang_map we were using doesn't actually work as spaces are not allowed in key:value pairs. We also had the wrong mapping for Brazilian so that might have caused issues down the line.