Always send experiment metadata with tracking events (#8024)
JIRA: https://soundcloud.atlassian.net/browse/DROID-1333

Previously we would only tag tracking events with experiments that the "app knows about" using the list from ActiveExperiments. However, now we are starting to drive experiments completely from the backend, where the client is totally agnostic about them. So now if roadie says the client is participating in an experiment, they are, and we will always send through the variants. For the situation where a client needs to do local UI/behaviour changes for an experiment, the experiment can be configured in roadie to _exclude_ older clients that don't have this logic implemented. This is the ticket that represents the work in Roadie to filter by a minimum app version: https://soundcloud.atlassian.net/browse/DSC-131