Expose ExperimentsHelper to control A/B test assignments (#8069)
In order to properly test our A/B test variations, we need a class
that would force the user assignment. ExperimentOperations
provides a way to write to SharedPreferences and thus manually
control the variant for each A/B test.

This is exposing this control into a class where we will be able to manually
set the assignment for the ExperimentConfigurations and thus enable us
to test every variant for our A/B tests.