Fix bug with some callers of ImageOperations not doing artwork fallback for playlists (#8071)
https://soundcloud.atlassian.net/browse/DROID-1344

When we refactored ImageOperations to take a cacheKey instead of a Urn, we accidentally
broke the mechanism that builds a fallback image url based on the Urn. This change brings back,
Urn, but it's optional, for the rare but real cases where we want to display an image for something
that doesn't have a Urn. If we have an image-url but not a Urn, then we will use the image-url as the cache-key.
If we have neither a image-url or a Urn, then we will display a default placeholder.

fix profile image loading and add NonNull annotation