Enable flipper based on a/b experiment or feature flag (#8148)
Instead of just using a feature flag, we decided to make an experiment to compare
listening time and #plays for each player (skippy / flipper).
This changes the current behaviour of the feature flag to be enabled by the default and
sets it to `DEFAULT_HIDE` in every flavour other than `debug`. The actual activation
will then be controlled by the `Experiment` class.

So, if a user is either in the proper `variant` or explicitly enabled the feature flag,
then flipper will be enabled.