Report table stats along with slow queries (#8231)
We receive lots of “slow query” silent exceptions. When debugging them, we currently only see which queries took a lot of time, but because of the nature of database locking and how we report the queries, it is often tough to pinpoint the exact cause of the problems.

This is an effort to get more info, by reporting table row counts from our main database along with the reports. Hopefully we can see some correlations then.