Always report slow queries, regardless of props (#8246)
We want to know if we still have slow queries, even in production. This changes it so that we do not have to enable query logging just to do this. It will always send the silent exception so we have number on this.

I made it so it outputs the last 30 queries right before it reports a slow query, so we didn’t have to log everything.