Perform scrub test only on long tracks (#8364)
When performing the scrubbing test, we were using the first reposted track.
If that is a snipped track or a short track (some seconds) scrubbing could
trigger skipping the track which would fail the test.
By searching for long tracks, we avoid the problem. I have introduced a new matcher for it and update the acceptance test.

PS: Notice how we use different IDs for the track card element in the likes screen (`R.id.list_item_right_info`) and in the stream (`R.id.duration`). Maybe we should work on aligning those.