Anchor snackbar on top most view from player hierarchy (#8525)
Since we added an overlay on top of the player bottom sheet, the mechanism for showing snackbars is wrongly
following the z-index of the views. Instead of anchoring snackbars on the player itself, let's anchor them at the
top-most window view (that is, `android.R.id.content`) to make sure they are rendered whatever the view hierarchy
is beneath.