Fix `MemoryCacheUtils` crash (#8650)
It was introduced in the 2017-09-13 beta, although that particular bit of the code didn't seem to change. It looks like we've always had the possibility to crash here as `imageUrlBuilder.buildUrl` can return null.

The method is nullable anyway so now we just return null when a url can't be created. Added a few @Nullable annotations to make it clearer as well.