Local search history (#8660)
Making a new PR as I've made some changes since #8447 (mostly ui and uniflow)

Attempted a very simple search history to save queries that you search for, as well as suggestions and autocompletions that you click on, to a local storage. It limits to 5 results and truncates the db, and you can clear it from the settings. It's behind a feature flag and only turned on for alpha.

The only other thing I need to do is move the clear button to the bottom of the list, which I didn't quite finish off on Friday, but I think it makes sense for people to have a play around with it sooner.