Return local current index for playerPagerPresenter (#8756)
This is an attempt to fix

https://soundcloud.atlassian.net/browse/DROID-1467

I could not reproduce it. It seems to be a race condition where we receive an invalid position from the TrackPager, which does not exist in our playqueue. The order of execution when setting a new playqueue is:

1. Set local current index
2. Set new queue 
3. Call notifyDataSetChanged (to update the pager)

Currently we have no guarantees around when the pager is ACTUALLY updated to reflect the new position. 

If it is synchronous, that means we are checking the position from another thread. I added an assertion around this that should crash us in alpha and below.

If it is not synchronous, that means we should default to the pager index, but fallback, as it may be invalid, and not set yet.