Treat policy as nullable when loading tracks from database since legacy tracks policy can be null (#8847)
Enforce non-nullability of track policies:
At write time, check that policy isn't null and if it is, throw an IllegalStateException.
At read time, check if policy is null, if it is, filter it out from whatever list it's showing up in and log a silent exception

Also includes a fix for the most likely candidate for what is causing this bad data: when uploading tracks, we post to public API then put the response in the database. Tracks from public API do not include policies. For now, treat the missing policy as ALLOW since we just uploaded this track and know what the policy should be.